+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add AppAsgKeyName for optional SSH access
 * Update descriptions and tags
+* Add option to restrict IP CIDR on ALB SG
 
 # 1.0.0